api/requirements.txt
fastapi>=0.110.0,<0.111.0
pydantic>=2.7.0,<2.8.0
pydantic[email]
uvicorn>=0.25.0,<0.26.0
python-dotenv
sqlalchemy
pydantic_settings
PyJWT